+++ Other/inversions.cpp
@@ -1,8 +1,8 @@
 template<typename T>
-ll merge_sort(vector<T>& v, vector<T>& aux, int l, int r) {
+long long merge_sort(vector<T>& v, vector<T>& aux, int l, int r) {
     if (r <= l) return 0LL;
     int m = l + (r - l) / 2;
-    ll res = merge_sort(v, aux, l, m) + merge_sort(v, aux, m + 1, r);
+    long long res = merge_sort(v, aux, l, m) + merge_sort(v, aux, m + 1, r);
     int x = l, y = m + 1, z = l;
     while (x <= m && y <= r) {
         aux[z++] = v[x] <= v[y] ? v[x++] : (res += m - x + 1, v[y++]);
@@ -13,7 +13,7 @@ ll merge_sort(vector<T>& v, vector<T>& aux, int l, int r) {
     return res;
 }
 template<typename T>
-ll inversions(vector<T> v) {
+long long inversions(vector<T> v) {
     static vector<T> aux;
     aux.resize(max(size(aux), size(v)));
     return merge_sort(v, aux, 0, (int)size(v) - 1);
